pyproject.toml
[project]
name = "repl-nix-llmconnector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.10",
    "anthropic>=0.37.1",
    "asyncio>=3.4.3",
    "google-generativeai>=0.8.3",
    "openai>=1.52.2",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "streamlit>=1.39.0",
]
